+++ packages/api/modules/admin/procedures/list-users.ts
@@ -12,8 +12,8 @@ export const listUsers = adminProcedure
 	.input(
 		z.object({
 			query: z.string().optional(),
-			limit: z.number().min(1).max(100).default(10),
-			offset: z.number().min(0).default(0),
+			limit: z.number().int().min(1).max(100).default(10),
+			offset: z.number().int().min(0).default(0),
 		}),
 	)
 	.handler(async ({ input: { query, limit, offset } }) => {
